# genai-observability-platform/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GenAI Observability Platform API
  description: |
    REST API for the GenAI Observability Platform.

    ## Authentication

    The API supports two authentication methods:

    1. **API Key** (for SDK/agents): Pass in `X-API-Key` header
    2. **JWT Token** (for dashboard): Pass in `Authorization: Bearer <token>` header

    ## Rate Limiting

    - API keys: 1000 requests/minute
    - JWT tokens: 100 requests/minute

    ## Pagination

    List endpoints support pagination via `limit` and `offset` query parameters.
    Response includes `total` count and `has_more` boolean.

  version: 1.0.0
  contact:
    name: Platform Team
    email: platform@example.com
  license:
    name: MIT
servers:
  - url: https://api.observability.example.com
    description: Production
  - url: https://api.staging.observability.example.com
    description: Staging
  - url: http://localhost:8000
    description: Local Development
...
paths:
...
  /api/v1/traces:
    get:
      tags: [Traces]
      summary: List traces
      operationId: listTraces
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: agent_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [success, error]
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of traces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/traces/{trace_id}:
    get:
      tags: [Traces]
      summary: Get trace details
      operationId: getTrace
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: trace_id
          in: path
          required: true
          schema:
            type: string
        - name: include_spans
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Trace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      operationId: listAlerts
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, info]
        - name: agent_id
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertList'
  /api/v1/alerts/{alert_id}:
    get:
      tags: [Alerts]
      summary: Get alert details
      operationId: getAlert
      security:
        - BearerAuth: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Alerts]
      summary: Update alert status
      operationId: updateAlert
      security:
        - BearerAuth: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertUpdate'
      responses:
        '200':
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
    TraceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TraceSummary'
        total:
          type: integer
        has_more:
          type: boolean
    TraceSummary:
      type: object
      properties:
        trace_id:
          type: string
        agent_id:
          type: string
        root_span_name:
          type: string
        start_time:
          type: string
          format: date-time
        duration_ms:
          type: integer
        status:
          type: string
          enum: [success, error]
        span_count:
          type: integer
    TraceDetail:
      allOf:
        - $ref: '#/components/schemas/TraceSummary'
        - type: object
          properties:
            spans:
              type: array
              items:
                $ref: '#/components/schemas/Span'
            total_tokens:
              type: integer
            total_cost_usd:
              type: number
            error:
              type: object
    Span:
      type: object
      properties:
        span_id:
          type: string
        parent_span_id:
          type: string
        name:
          type: string
        span_type:
          type: string
        start_time:
          type: string
          format: date-time
        duration_ms:
          type: integer
        status:
          type: string
        attributes:
          type: object
        error:
          type: object
...
    AlertList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        total:
          type: integer
    Alert:
      type: object
      properties:
        alert_id:
          type: string
          format: uuid
        rule_id:
          type: string
          format: uuid
        agent_id:
          type: string
        trace_id:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, acknowledged, resolved]
        triggered_at:
          type: string
          format: date-time
        metrics:
          type: object
        investigation:
          type: object
    AlertUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [acknowledged, resolved]
        note:
          type: string
        resolution_note:
          type: string
...